+++ .gitignore
@@ -44,3 +44,5 @@ buck-out/
 \.buckd/
 *.keystore
       
+# Web build
+web/
