+++ package.json
@@ -2,7 +2,12 @@
   "name": "@react-native-community/geolocation",
   "version": "1.3.0",
   "description": "React Native Geolocation Module for iOS and Android",
-  "main": "js/index.js",
+  "react-native": "js/index.js",
+  "main": "web/commonjs/index.js",
+  "files": [
+    "web/",
+    "js/"
+  ],
   "typings": "typings/index.d.ts",
   "author": "React Native Community",
   "contributors": [
@@ -25,7 +30,8 @@
     "build:e2e:ios:debug": "detox build -c ios.sim.debug",
     "build:e2e:ios:release": "detox build -c ios.sim.release",
     "test:e2e:ios:debug": "detox test -c ios.sim.debug",
-    "test:e2e:ios:release": "detox test -c ios.sim.release"
+    "test:e2e:ios:release": "detox test -c ios.sim.release",
+    "prepare": "bob build"
   },
   "keywords": [
     "react-native",
@@ -92,5 +98,16 @@
       }
     },
     "specs": ""
+  },
+  "dependencies": {
+    "@react-native-community/bob": "^0.3.4"
+  },
+  "@react-native-community/bob": {
+    "source": "js",
+    "output": "web",
+    "targets": [
+      ["commonjs", {"flow": true}],
+      "module"
+    ]
   }
 }
